docker-assignment/solution/requirements.txt
flask
psycopg2-binary==2.8.4
